Working on DOGM code...
