+++ Marlin/ultralcd_impl_HD44780.cpp
@@ -1,6 +1,6 @@
+#include "Marlin.h"
 #include "ultralcd.h"
 #if ENABLED(ULTRA_LCD)
-#include "Marlin.h"
 #include "language.h"
 #include "cardreader.h"
 #include "temperature.h"
